int to long

+++ src/wmtk/operations/tri_mesh/BuildOffset.hpp
@@ -15,7 +15,7 @@ struct OperationSettings<tri_mesh::BuildOffset>
     // handle to vertex position
     MeshAttributeHandle<double> position;
     // handle to vertex tag
-    MeshAttributeHandle<int> tag;
+    MeshAttributeHandle<long> tag;
 
     int pass = -1;
 
@@ -44,14 +44,14 @@ protected:
 private:
     Tuple m_output_tuple;
     Accessor<double> m_pos_accessor;
-    Accessor<int> m_tag_accessor; // 0-scalffold 1-input 2-offset
+    Accessor<long> m_tag_accessor; // 0-scalffold 1-input 2-offset
 
     const OperationSettings<BuildOffset>& m_settings;
 
     Eigen::Vector3d p0;
     Eigen::Vector3d p1;
-    int t0;
-    int t1;
+    long t0;
+    long t1;
 };
 
 } // namespace tri_mesh
